pyproject.toml
[project]
name = "xuan-flow"
version = "0.1.0"
description = "A lightweight LangGraph-based Multi-Agent framework inspired by DeerFlow"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
